requirements.txt
atlassian-python-api
python-dotenv
lxml
inquirer
requests
